Bugfixing super refraction filters (#3199)
* clean up; remove 'Variables observed'

* check if geoval variables are missing

* formatting

* function type buf fix

* check if the profile does not penetrate below the maxCheckHeight

* comment

* tidy up

* remove unnecessary 'clear'

* edit comments

* remove unnecessary clear

* kick off full ci check after taking this out of draft

* fix comments

* minor change to NBAM filter: remove clear

---------

+++ src/ufo/filters/ProfileMaxDifferenceCheck.h
@@ -71,9 +71,9 @@ class ProfileMaxDifferenceCheck : public FilterBase,
                    std::vector<std::vector<bool>> &) const override;
   int qcFlag() const override {return QCflags::profile;}
   Parameters_ parameters_;
-  float calcMaxDifference(const std::vector<float> &,
-                          const std::vector<float> &)
-                          const;
+  int calcMaxDifference(const std::vector<float> &,
+                        const std::vector<float> &)
+                        const;
 };
 
 }  // namespace ufo
